router: use HashRouter for routing
The previous BrowserRouter did not work when reloading the page, due to
how it is implemented.

HashRouter is discouraged since it doesn't work for SEO, however SEO is
the last thing we need on a Farpatch device -- if a Farpatch device
shows up in a search engine, it's Very Bad.

Use the HashRouter to avoid having to adjust the http server running on
Farpatch to work with our router.

Signed-off-by: Sean Cross <sean@xobs.io>